A_Vihave-Yolo-Dashboard-v1.8/A_Vihave-Yolo-Dashboard/requirements.txt
absl-py 
aiocontextvars==0.2.2
altair==4.1.0
altgraph==0.17.2
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
azure-ai-formrecognizer==3.1.2
azure-common==1.1.28
azure-core==1.24.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.4
cairocffi==1.2.0
CairoSVG==2.5.2
camelot-py==0.10.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.10
ChatterBot 
clang==5.0
click==8.0.3
colorama==0.4.4
config==0.5.1
contextlib2==21.6.0
contextvars==2.4
cryptography==36.0.2
cssselect2==0.4.1
cvzone==1.5.6
cycler==0.11.0
Cython==0.29.30
dash==2.5.1
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-extensions==0.0.71
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses 
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
djitellopy==2.4.0
dominate==2.6.0
easydict==1.9
EditorConfig==0.12.3
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.75.0
filelock==3.4.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Compress==1.12
Flask-Cors==3.0.7
flatbuffers==1.12
future==0.18.2
gast==0.4.0
ghostscript==0.7
gitdb==4.0.9
GitPython==3.1.20
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.1.0
html5lib==1.1
huggingface-hub==0.4.0
idna==3.3
imageio==2.15.0
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipdb==0.13.9
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.0.1
jarowinkler==1.0.2
jedi==0.17.2
Jinja2==2.11.3
joblib==1.1.0
jsbeautifier==1.14.4
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.6.0
Keras-Applications
Keras-Preprocessing 
kiwisolver==1.3.1
loguru==0.6.0
Markdown==3.3.6
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.13.0
MouseInfo==0.1.3
mplcursors==0.5.1
msrest==0.7.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
ninja==1.10.2.3
nltk==3.6.7
notebook==6.4.10
numexpr==2.8.1
numpy 
oauthlib==3.2.0
onnx==1.5.0
onnx-simplifier==0.3.5
onnxoptimizer==0.2.7
onnxruntime==1.8.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pafy==0.5.5
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pdfplumber==0.6.0
pdftabextract==0.3.0
pdftopng==0.2.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.9.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pyclipper==1.3.0.post2
pycocotools==2.0.4
pycparser==2.21
pydantic==1.9.0
pydeck==0.6.2
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.7
PyPDF2==1.27.2
pypdfium2==2.0.0
pyperclip==1.8.2
pyphen==0.11.0
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
pyxlsb==1.0.9
PyYAML==6.0
pyzmq==22.3.0
rapidfuzz==2.0.11
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.2
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
starlette==0.17.1
streamlit==1.7.0
tables==3.7.0
tabula-py==2.3.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-cpu==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tf2onnx==1.11.1
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2020.9.3
tinycss2==1.1.1
tk==0.1.0
toml==0.10.2
toolz==0.11.2
torch==1.10.1
torchsummary==1.5.1
tzdata==2022.1
tzlocal==4.1
Unidecode==1.3.4
urllib3==1.26.7
uvicorn==0.16.0
validators==0.18.2
visitor==0.1.3
Wand==0.6.7
watchdog==2.1.6
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug
what3words==3.1.1
widgetsnbextension==3.6.0
win32-setctime==1.1.0
wincertstore==0.2
wrapt==1.12.1
XlsxWriter==3.0.3
zipp 
